Added base class for subtitle tests
