+++ PySubtitle/UnitTests/test_Project.py
@@ -1,6 +1,5 @@
-import unittest
-
-from PySubtitle.Helpers.Tests import PrepareSubtitles, log_info, log_input_expected_result, log_test_name
+from PySubtitle.Helpers.TestCases import PrepareSubtitles, SubtitleTestCase
+from PySubtitle.Helpers.Tests import log_info, log_input_expected_result, log_test_name
 from PySubtitle.Options import Options
 from PySubtitle.SubtitleBatch import SubtitleBatch
 from PySubtitle.SubtitleBatcher import SubtitleBatcher
@@ -10,7 +9,7 @@ from PySubtitle.SubtitleProject import SubtitleProject
 from PySubtitle.SubtitleScene import SubtitleScene
 from PySubtitle.UnitTests.TestData.chinese_dinner import chinese_dinner_data
 
-class ChineseDinnerTests(unittest.TestCase):
+class ChineseDinnerTests(SubtitleTestCase):
     options = Options({
         'target_language': 'English',
         'scene_threshold': 60.0,
@@ -85,7 +84,7 @@ class ChineseDinnerTests(unittest.TestCase):
         ]
         batch_containing_line = [(1, 1, 1), (10, 1,1), (32, 2, 1), (55, 2, 1), (63, 4, 1)]
 
-        subtitles = PrepareSubtitles(chinese_dinner_data)
+        subtitles : SubtitleFile = PrepareSubtitles(chinese_dinner_data)
 
         batcher = SubtitleBatcher(self.options)
         subtitles.AutoBatch(batcher)
